+++ calculator.h
@@ -8,15 +8,15 @@
 template <typename Operator>
 class Calculator {
     private:
-        int current_result = 0;
-        int saved_result = 0;
-        Operator *op = [](int b, int c){return b;};
+        int64_t current_result = 0;
+        int64_t saved_result = 0;
+        Operator *op = [](int64_t b, int64_t c){return b;};
     public:
         Calculator() {};
         void addDigit(uint8_t digit);
         void useOperator();
         void setOperator(const Operator* func);
-        int getResult();
+        int64_t getResult();
 };
 
 #endif
